requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
pymupdf==1.23.8
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
langchain==0.0.340
unstructured==0.10.30
nltk==3.8.1
opencv-python==4.8.1.78
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4